Rename folders to match naming convention
