+++ src/BigQueryFunctions/CreateIssueTable.ts
