+++ src/BigQueryFunctions/DeleteTable.ts
